SQL vulnerability update for DBController class
NOTE: This vulnerability was intentionally added for a school project as part of a software security class.

All methods that used prepared statements now use string concatenation instead.

If this update causes you to encounter issues elsewhere in the project, please contact me and I'll address them as soon as possible.